third_party/WebKit/LayoutTests/http/tests/media/video-controls-download-button-not-displayed-hide-download-ui.html

<!DOCTYPE HTML>
<title>Download button should not appear when hiding download UI.</title>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="../../media-resources/media-controls.js"></script>
<video controls></video>
<script>
async_test(function(t) {
  internals.settings.setHideDownloadUI(true);
  var video = document.querySelector("video");
  video.src = "resources/test.ogv";
  var controlID = "-internal-media-controls-download-button";
  var downloadButton = mediaControlsElement(internals.shadowRoot(video).firstChild, controlID);
  video.onloadeddata = t.step_func_done(function() {
      assert_equals(getComputedStyle(downloadButton).display, "none");
  });
});
</script>
